🧪 Improve testing coverage for SurveyTranslationManager
This commit adds testing to SurveyTranslationManager and OpenAiTranslationClient by:
- Fixing the location of the obsolete `lite/survey` directory to `lite/surveys` while preserving duplicate test configurations.
- Adding tests for `OpenAiTranslationClient` for successful HTTP responses and input validations.
